+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-<p align="center">
-    <img src="https://cdn.marshmallow-office.com/media/images/logo/marshmallow.transparent.red.png">
-</p>
+![marshmallow-transparent-logo](https://cdn.marshmallow-office.com/media/images/logo/marshmallow.transparent.red.png)(https://marshmallow.dev)
 
 # Change Log
 
+## 1.3.0-beta - 25-April-2020
+
+- Updated Publishing.
+- Added Responsive Theme & Options.
 
 ## 1.0.0-beta - 25-Jan-2020